Move `MinMax` type into mz-ore
This commit moves the `MinMax` type supporting the compute controller
metrics into the mz-ore crate, as it might be generally useful. The type
is also renamed to `SlidingMinMax` to make its purpose more obvious
outside the a metrics context.
