+++ src/compute-client/src/metrics.rs
@@ -22,7 +22,7 @@ use mz_ore::metrics::{
     CounterVec, DeleteOnDropCounter, DeleteOnDropGauge, DeleteOnDropHistogram, GaugeVec,
     HistogramVec, IntCounterVec, MetricVecExt, MetricsRegistry,
 };
-use mz_ore::stats::histogram_seconds_buckets;
+use mz_ore::stats::{histogram_seconds_buckets, SlidingMinMax};
 use mz_repr::GlobalId;
 use mz_service::codec::StatsCollector;
 use prometheus::core::{AtomicF64, AtomicU64};
@@ -436,7 +436,7 @@ impl ReplicaMetrics {
             .metrics
             .dataflow_wallclock_lag_seconds_count
             .get_delete_on_drop_metric(labels);
-        let wallclock_lag_minmax = MinMax::new(60);
+        let wallclock_lag_minmax = SlidingMinMax::new(60);
 
         Some(ReplicaCollectionMetrics {
             initial_output_duration_seconds,
@@ -483,7 +483,7 @@ pub(crate) struct ReplicaCollectionMetrics {
     wallclock_lag_seconds_count: IntCounter,
 
     /// State maintaining minimum and maximum wallclock lag.
-    wallclock_lag_minmax: MinMax<f32>,
+    wallclock_lag_minmax: SlidingMinMax<f32>,
 }
 
 impl ReplicaCollectionMetrics {
@@ -501,129 +501,6 @@ impl ReplicaCollectionMetrics {
         self.wallclock_lag_seconds_max.set(max.into());
         self.wallclock_lag_seconds_sum.inc_by(lag_secs.into());
         self.wallclock_lag_seconds_count.inc();
-    }
-}
-
-/// Keeps track of a the minimum and maximum value over a fixed-size sliding window of samples.
-///
-/// Inspired by the [`movning_min_max`](https://crates.io/crates/moving_min_max) crate, but
-/// adapted to provide:
-///
-///  * both minimum and maximum at the same time
-///  * a fixed window size, rather than manual popping of samples
-///  * lower memory usage, due to merged push and pop stacks
-#[derive(Debug)]
-struct MinMax<T> {
-    /// The push stack and the pop stack, merged into one allocation.
-    ///
-    /// The push stack is the first `push_stack_len` items, the pop stack is the rest.
-    /// The push stack grows to the left, the pop stack grows to the right.
-    ///
-    /// +--------------------------------+
-    /// | push stack --> | <-- pop stack |
-    /// +----------------^---------------^
-    ///           push_stack_len
-    ///
-    /// We keep both the push stack and the pop stack in the same `Vec`, to guarantee that we only
-    /// allocate `size_of::<T>() * 3 * window_size` bytes. If we kept two `Vec`s we'd need twice
-    /// that much.
-    stacks: Vec<(T, T, T)>,
-    /// The length of the push stack.
-    push_stack_len: usize,
-}
-
-impl<T> MinMax<T>
-where
-    T: Clone + PartialOrd,
-{
-    /// Creates a new `MinMax` for the given sample window size.
-    fn new(window_size: usize) -> Self {
-        Self {
-            stacks: Vec::with_capacity(window_size),
-            push_stack_len: 0,
-        }
-    }
-
-    /// Returns a reference to the item at the top of the push stack.
-    fn top_of_push_stack(&self) -> Option<&(T, T, T)> {
-        self.push_stack_len.checked_sub(1).map(|i| &self.stacks[i])
-    }
-
-    /// Returns a reference to the item at the top of the pop stack.
-    fn top_of_pop_stack(&self) -> Option<&(T, T, T)> {
-        self.stacks.get(self.push_stack_len)
-    }
-
-    /// Adds the given sample.
-    fn add_sample(&mut self, sample: T) {
-        if self.push_stack_len == self.stacks.capacity() {
-            self.flip_stacks();
-        }
-
-        let (min, max) = match self.top_of_push_stack() {
-            Some((_, min, max)) => {
-                let min = po_min(min, &sample).clone();
-                let max = po_max(max, &sample).clone();
-                (min, max)
-            }
-            None => (sample.clone(), sample.clone()),
-        };
-
-        if self.stacks.len() <= self.push_stack_len {
-            self.stacks.push((sample, min, max));
-        } else {
-            self.stacks[self.push_stack_len] = (sample, min, max);
-        }
-        self.push_stack_len += 1;
-    }
-
-    /// Drains the push stack into the pop stack.
-    fn flip_stacks(&mut self) {
-        let Some((sample, _, _)) = self.top_of_push_stack().cloned() else {
-            return;
-        };
-
-        self.push_stack_len -= 1;
-        self.stacks[self.push_stack_len] = (sample.clone(), sample.clone(), sample);
-
-        while let Some((sample, _, _)) = self.top_of_push_stack() {
-            let (_, min, max) = self.top_of_pop_stack().expect("pop stack not empty");
-            let sample = sample.clone();
-            let min = po_min(min, &sample).clone();
-            let max = po_max(max, &sample).clone();
-
-            self.push_stack_len -= 1;
-            self.stacks[self.push_stack_len] = (sample, min, max);
-        }
-    }
-
-    /// Returns the current minimum and maximum values.
-    fn get(&self) -> Option<(&T, &T)> {
-        match (self.top_of_push_stack(), self.top_of_pop_stack()) {
-            (None, None) => None,
-            (None, Some((_, min, max))) | (Some((_, min, max)), None) => Some((min, max)),
-            (Some((_, min1, max1)), Some((_, min2, max2))) => {
-                Some((po_min(min1, min2), po_max(max1, max2)))
-            }
-        }
-    }
-}
-
-/// Like `std::cmp::min`, but works with `PartialOrd` values.
-fn po_min<T: PartialOrd>(a: T, b: T) -> T {
-    if a < b {
-        a
-    } else {
-        b
-    }
-}
-
-/// Like `std::cmp::max`, but works with `PartialOrd` values.
-fn po_max<T: PartialOrd>(a: T, b: T) -> T {
-    if a > b {
-        a
-    } else {
-        b
     }
 }
 
@@ -808,36 +685,5 @@ impl<M> PeekMetrics<M> {
             Error(_) => &self.error,
             Canceled => &self.canceled,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[mz_ore::test]
-    fn minmax() {
-        let mut minmax = MinMax::new(5);
-
-        assert_eq!(minmax.get(), None);
-
-        let mut push_and_check = |x, expected| {
-            minmax.add_sample(x);
-            let actual = minmax.get().map(|(min, max)| (*min, *max));
-            assert_eq!(actual, Some(expected), "{minmax:?}");
-        };
-
-        push_and_check(5, (5, 5));
-        push_and_check(1, (1, 5));
-        push_and_check(10, (1, 10));
-        push_and_check(2, (1, 10));
-        push_and_check(9, (1, 10));
-        push_and_check(3, (1, 10));
-        push_and_check(8, (2, 10));
-        push_and_check(5, (2, 9));
-        push_and_check(5, (3, 9));
-        push_and_check(5, (3, 8));
-        push_and_check(5, (5, 8));
-        push_and_check(5, (5, 5));
     }
 }
